backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

sqlalchemy>=2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
boto3==1.34.0
pydantic>=2.5.3
pydantic-settings>=2.1.0
aiofiles==23.2.1
emails==0.6
jinja2==3.1.3
httpx==0.26.0
requests==2.31.0
